requirements.txt
Python 3.13.9,
fastapi  0.121.3
uvicorn  0.38.0
pydantic 2.12.4
python-dateutil #python-dateutil is a powerful date parsing library.
boto3    #Use boto3 later for production
openai 2.14.0
python-dotenv 1.2.1
pytest 9.0.2   # pytest -v